Move mocks to __tests__ folder
